Fix DLL extraction/download for Windows
It seems like we are not allowed to rename the file, so we need to create a temporary directory and put it in there preserving its original name.

+++ src/main/java/de/jcm/discord/activity/Launcher.java
@@ -37,7 +37,12 @@ public class Launcher
 		{
 			if(entry.getName().equals(zipPath))
 			{
-				File temp = File.createTempFile(name, suffix);
+				File tempDir = new File(System.getProperty("java.io.tmpdir"), "java-"+name+System.nanoTime());
+				if(!tempDir.mkdir())
+					throw new IOException("Cannot create temporary directory");
+				tempDir.deleteOnExit();
+
+				File temp = new File(tempDir, name+suffix);
 				temp.deleteOnExit();
 
 				FileOutputStream fout = new FileOutputStream(temp);
